relocate composed function outside component
